First commit!
This is an implementation of that one paper that is really awesome and
has the formulas for "updating" sets of stats.  By "updating" the paper
means supporting the addition operation on "views" of datasets consisting
of the central moments.  So that liek, adding the two datasets and then
calculating the moments would have the same result as calculating the
moments for each dataset and adding the "views".  There's some very
simple word for this that you can learn in discrete math or abstract
algebra, but I don't remember it right now.

Also, it is numerically stable, so it won't produce wildly incorrect
stddev like the naive implementation.

There don't really need to be more commits, but maybe I should add a
reference to the really awesome paper once I find it again, and also a
README so that people will READ and UNDERSTAND.